documentation save find

+++ src/main/kotlin/com/ryouonritsu/inkbook_backend/dao/DocumentationDao.kt
@@ -10,5 +10,5 @@ interface DocumentationDao {
     fun update(documentation: Documentation)
     fun find(doc_id: Long): Documentation?
     fun findByProjectId(project_id: Int): List<Documentation>
-    fun findByCreatorId(creator_id: Int): List<Documentation>
+    fun findByCreatorId(creator_id: Long): List<Documentation>
 }
